Only accept a single library for each name, and move SystemPath out of Binaries
Without this check in place, provider preferences were actually quite broken; the least preferred would get picked, and hooks would get mixed with library paths from other providers.  See https://github.com/staticfloat/homebrew-juliadeps/issues/38 for an example.

Moving SystemPath out of Binaries is necessary so that it is not included in `defaults` more than once, and so it has the proper ordering in preference